Made a case for when all of the samples PASS mean target coverage and reconfigured the test data folder.
